Refactor MenuUI for significantly improved clarity
